+++ nodes/RosTopicNextMessage/RosTopicNextMessage.node.ts
@@ -207,7 +207,6 @@ export class RosTopicNextMessage implements INodeType {
     async execute(this: IExecuteFunctions): Promise<INodeExecutionData[][]> {
         const items = this.getInputData();
         const returnData: INodeExecutionData[] = [];
-        const continueOnFail = this.continueOnFail();
 
         for (let i = 0; i < items.length; i++) {
             let ros;
@@ -253,7 +252,7 @@ export class RosTopicNextMessage implements INodeType {
                     pairedItem: { item: i },
                 });
             } catch (error) {
-                if (continueOnFail) {
+                if (NodeErrorHandler.shouldReturnErrorOutput(this)) {
                     returnData.push({
                         json: NodeErrorHandler.buildErrorOutput(error as Error),
                         pairedItem: { item: i },